# swagger.yaml
openapi: 3.0.0
info:
  title: raank/xml-reader
  description: 'This is a micro authentication service and users or files crud. '
  contact:
    email: raank92@gmail.com
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8081/api'
paths:
...
  /v1/files:
    post:
      tags:
        - v1.files
      summary: 'Storing a new File to User.'
      operationId: f48c68b8dab381c4d0e995b60c94b6e4
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
                - user_id
              properties:
                file:
                  description: 'The file to parse.'
                user_id:
                  description: 'The user id to relation.'
                  type: integer
                  example: 1
              type: object
      responses:
        '202':
          description: 'Successful action'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Message of Response', type: string, example: 'Successful action' }
                  data: { $ref: '#/components/schemas/v1.models.file' }
                type: object
        '400':
          description: 'This information could not be processed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 'You are not authorized for this action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
      security:
        -
          apiToken: []
  '/v1/files/{fileId}':
    get:
      tags:
        - v1.files
      summary: 'Show specified file.'
      operationId: e09671ff48d4f0de1ae623e8405c98bd
      parameters:
        -
          name: fileId
          in: path
          description: 'Identification of File'
          required: true
          schema:
            type: integer
          example: 2
      responses:
        '200':
          description: 'Successful action'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Message of Response', type: string, example: 'Successful action' }
                  data: { $ref: '#/components/schemas/v1.models.file' }
                type: object
        '400':
          description: 'This information could not be processed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 'You are not authorized for this action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 'This information could not be found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notfound'
      security:
        -
          apiToken: []
    delete:
      tags:
        - v1.files
      summary: 'Delete specified file.'
      operationId: 27318961b2a6fb666adb4ae79244444f
      parameters:
        -
          name: fileId
          in: path
          description: 'Identification of File'
          required: true
          schema:
            type: integer
          example: 2
      responses:
        '200':
          description: 'Successful action'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Message of Response', type: string, example: 'Successful action' }
                type: object
        '400':
          description: 'This information could not be processed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 'You are not authorized for this action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 'This information could not be found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notfound'
      security:
        -
          apiToken: []
  /v1/files/search:
    post:
      tags:
        - v1.files
      summary: 'Searching files.'
      operationId: bdbb158604c34a4f6446c59a94dd9511
      requestBody:
        content:
          application/json:
            schema:
              properties:
                where:
                  description: 'Where Condition.'
                whereNotNull:
                  description: 'Where field not null.'
                whereNull:
                  description: 'Where field is nullable.'
                orderBy:
                  description: 'The password confirmation.'
                whereBetween:
                  description: 'The filename of file.'
              type: object
              example:
                where:
                  - [field_name, operator, value]
                  - [name, LIKE, john]
                whereNotNull:
                  - field_name
                whereNull:
                  - field_name
                orderBy:
                  - { field_name: field_name, order: DESC }
                whereBetween:
                  - { field_name: [from_value, to_value] }
      responses:
        '200':
          description: 'Successful action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.paginated.files'
        '400':
          description: 'This information could not be processed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 'You are not authorized for this action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
      security:
        -
          apiToken: []
...
components:
  schemas:
    BadRequest:
      description: 'This information could not be processed'
      properties:
        message:
          description: 'Message of Response'
          type: string
          example: 'This information could not be processed'
      type: object
    Notfound:
      description: 'This information could not be found'
      properties:
        message:
          description: 'Message of Response'
          type: string
          example: 'This information could not be found'
      type: object
    Unauthorized:
      description: 'You are not authorized for this action'
      properties:
        message:
          description: 'Message of Response'
          type: string
          example: 'You are not authorized for this action'
      type: object
...
    v1.pagination.links:
      description: 'List of Links'
      properties:
        url:
          description: 'URL of Link'
          type: string
          example: 'http://localhost'
        label:
          description: 'Label of Link'
          type: string
          example: my-label
        active:
          description: 'Link is active'
          type: boolean
          example: true
      type: object
...
    v1.paginated.files:
      description: 'Response CRUD paginated'
      allOf:
        -
          $ref: '#/components/schemas/v1.pagination.links'
        -
          properties:
            message:
              description: 'Message of Response'
              type: string
              example: 'Successful action'
            data:
              type: array
              items:
                $ref: '#/components/schemas/v1.models.file'
            current_page:
              description: 'Current page'
              type: integer
              example: 1
            first_page_url:
              description: 'First page URL'
              type: string
              example: 'http://localhost'
            from:
              description: 'From start items'
              type: integer
              example: 1
            last_page:
              description: 'Last Page Number'
              type: integer
              example: 1
            last_page_url:
              description: 'Last page URL'
              type: string
              example: 'http://localhost'
            links:
              description: 'List of Links'
              type: array
              items:
                $ref: '#/components/schemas/v1.pagination.links'
            next_page_url:
              description: 'Next page URL'
              type: string
              example: 'http://localhost'
            path:
              description: 'Path of current URL'
              type: string
              example: 'http://localhost'
            per_page:
              description: 'Items number per page'
              type: integer
              example: 1
            prev_page_url:
              description: 'Prev page URL'
              type: string
              example: 'http://localhost'
            to:
              description: 'Items to end page'
              type: integer
              example: 1
            total:
              description: 'Total of Items'
              type: integer
              example: 1
          type: object
...
    v1.models.file:
      description: 'Response data of user'
      properties:
        id:
          description: 'Identification of User'
          type: integer
          example: 1
        original:
          description: 'The file original name.'
          type: string
          example: my-file.xml
        name:
          description: 'The file name.'
          type: string
          example: file_hashed.xml
        path:
          description: 'Email of User'
          type: string
          example: /path/to/file_hashed.xml
        size:
          description: 'The file size'
          type: integer
          example: 1234
        mimeType:
          description: 'The file mime type'
          type: string
          example: application/xml
        content:
          description: 'The content of file'
          type: object
        updated_at:
          description: 'Date of last updated'
          type: string
          example: '2021-01-01T00:00:00.000000Z'
        created_at:
          description: 'Date of Created'
          type: string
          example: '2021-01-01T00:00:00.000000Z'
        deleted_at:
          description: 'Date of Destroy'
          type: string
          example: null
      type: object
...
  securitySchemes:
    apiToken:
      type: apiKey
      name: X-App-Token
      in: header
    bearerAuth:
      type: http
      description: 'Informations of API.'
      scheme: bearer
...